fix namespace suggestion error on context switch
